requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
openai==1.3.7
supabase==2.0.0
pgvector==0.2.4
python-dotenv==1.0.0
annotated-types==0.6.0
anyio==3.7.1
blinker==1.7.0
certifi==2023.11.17
click==8.1.7
distro==1.8.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
jinja2==3.1.2
markupsafe==2.1.3
pydantic==2.5.2
pydantic-core==2.14.5
sniffio==1.3.0
tqdm==4.66.1
typing-extensions==4.8.0
werkzeug==3.0.1